Share `FiberMonitor` between JS and Native
