+++ core/js/src/main/scala/cats/effect/unsafe/BatchingMacrotaskExecutor.scala
@@ -40,7 +40,8 @@ import scala.util.control.NonFatal
 private[effect] final class BatchingMacrotaskExecutor(
     batchSize: Int,
     reportFailure0: Throwable => Unit
-) extends ExecutionContextExecutor {
+) extends ExecutionContextExecutor
+    with FiberExecutor {
 
   private[this] val queueMicrotask: js.Function1[js.Function0[Any], Any] =
     if (js.typeOf(js.Dynamic.global.queueMicrotask) == "function")
